Move TitleRestorer to ProjectFileManager
