+++ src/ProjectFileManager.h
@@ -135,4 +135,17 @@ private:
    bool mMenuClose{ false };
 };
 
+class wxTopLevelWindow;
+
+//! TitleRestorer restores project window titles to what they were,
+//! in its destructor.
+class TitleRestorer{
+public:
+   TitleRestorer( wxTopLevelWindow &window, AudacityProject &project );
+   ~TitleRestorer();
+   wxString sProjNumber;
+   wxString sProjName;
+   size_t UnnamedCount;
+};
+
 #endif
